Fix spurious Casablanca_DoubleJsonParsing failures
Sometimes when generating a random double in the
Casablanca_DoubleJsonParsing test, a whole number like 0 or 1 would be
returned. When DoubleProperty with a whole number was subsequently
serialized and parsed, it was not considered a double, causing a test
failure.

Now, there's an explicit check to see whether the generated number has a
fractional part to make the right assertion.